visionctl/scripts/pyproject.toml
[project]
name = "visionctl-detection"
version = "0.1.0"
description = "Visual detection scripts for visionctl"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.8",
    "numpy>=1.26",
    "pillow>=10.0",
]

[project.optional-dependencies]
ml = [
    "ultralytics>=8.0",  # YOLOE/YOLO-World
]
grounding = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.30",
    "accelerate>=0.20",
]

[tool.uv]
dev-dependencies = []
